Work ongoing with the reminder feature.
This commit majorly adds the reminder page to the
frontend and the handler for the reminder page in
the server. Others are just refacotring and minor
changes to the codebase.
The reminder includes the "create reminder" modal
dialog, works still ongoing on the UI feature.
The backend endpoint is yet to be wired up, on next
commit.

Signed-off-by: tobigiwa <giwaoluwatobi@gmail.com>